backend/requirements.txt
streamlit
pandas
pandasql
python-dotenv
google-genai
SQLAlchemy
psycopg2-binary
langchain-pinecone
langchain-community